+++ .version
@@ -1,1 +1,1 @@
-0.6.4
+1.0.0
