+++ app/config/formatters.py
@@ -1,13 +1,13 @@
-from app.mappers.formatter_selectors import FormatterSelector
-from app.formatters.cora_formatter import CORAFormatter, MESFormatter
-from app.formatters.cora_looping_formatter import CORALoopingFormatter
-from app.formatters.cs_formatter import CSFormatter
-from app.formatters.cs_looping_formatter import CSLoopingFormatter
-from app.formatters.idbr_looping_formatter import IDBRLoopingFormatter
-from app.formatters.json_formatter import JSONFormatter
-from app.formatters.open_road_formatter import OpenRoadFormatter
-from app.formatters.spp_formatter import SPPFormatter
-from app.formatters.spp_looping_formatter import SPPLoopingFormatter
+from app.services.mappers.formatter_selectors import FormatterSelector
+from app.services.formatters.cora_formatter import CORAFormatter, MESFormatter
+from app.services.formatters.cora_looping_formatter import CORALoopingFormatter
+from app.services.formatters.cs_formatter import CSFormatter
+from app.services.formatters.cs_looping_formatter import CSLoopingFormatter
+from app.services.formatters.idbr_looping_formatter import IDBRLoopingFormatter
+from app.services.formatters.json_formatter import JSONFormatter
+from app.services.formatters.open_road_formatter import OpenRoadFormatter
+from app.services.formatters.spp_formatter import SPPFormatter
+from app.services.formatters.spp_looping_formatter import SPPLoopingFormatter
 
 
 _formatter_mapping = {
