+++ app/services/formatters/cora_formatter.py
@@ -1,5 +1,5 @@
 from app.definitions.data import SurveyMetadata, Empty, Value
-from app.formatters.formatter import Formatter
+from app.services.formatters.formatter import Formatter
 
 
 class CORAFormatter(Formatter):
