+++ app/services/formatters/spp_formatter.py
@@ -1,7 +1,7 @@
 import json
 
 from app.definitions.data import SurveyMetadata, PCK, SPP, Empty, Value
-from app.formatters.formatter import Formatter
+from app.services.formatters.formatter import Formatter
 
 
 class SPPFormatter(Formatter):
