+++ app/transformers/looped.py
@@ -2,7 +2,7 @@ from typing import Literal
 
 from app.definitions.data import SurveyMetadata
 from app.definitions.spec import BuildSpec, ParseTree
-from app.formatters.looping_formatter import LoopingFormatter
+from app.services.formatters.looping_formatter import LoopingFormatter
 from app.services.transform.populate import resolve_value_fields
 from app.transformers.standard import SpecTransformer
