--- requirements.txt
@@ -1,2 +1,0 @@
-sdx-gcp==1.4.1
-PyYAML~=6.0.1
