--- test-requirements.txt
@@ -1,4 +1,0 @@
-pytest==7.3.1
-pytest-cov==4.0.0
-flake8==6.0.0
-mypy==1.5.0
