+++ tests/unit/formatters/test_cora_looping_formatter.py
@@ -1,7 +1,7 @@
 import unittest
 
 from app.definitions.data import SurveyMetadata, PCK
-from app.formatters.cora_looping_formatter import CORALoopingFormatter
+from app.services.formatters.cora_looping_formatter import CORALoopingFormatter
 from tests.integration.flat import are_equal
